requirements.txt
Django==3.2.6
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
mysqlclient==2.0.3